+++ src/lib/arithmetic/index.ts
@@ -1,1 +1,1 @@
-export * from './change-date';
+export * from './add-date';